Add validator
