--- lib/elixir_google_scraper/scraper/csv_parser.ex
@@ -1,19 +1,0 @@
-defmodule ElixirGoogleScraper.Scraper.CSVParser do
-  alias NimbleCSV.RFC4180, as: CSV
-
-  def parse(%Plug.Upload{content_type: "text/csv"} = file) do
-    keyword_list =
-      file.path
-      |> File.stream!()
-      |> CSV.parse_stream(skip_headers: false)
-      |> Enum.to_list()
-
-    if length(keyword_list) > 0 do
-      {:ok, keyword_list}
-    else
-      {:error, :file_is_empty}
-    end
-  end
-
-  def parse(_), do: {:error, :file_is_invalid}
-end
