+++ lib/elixir_google_scraper_web/controllers/keyword_controller.ex
@@ -13,13 +13,21 @@ defmodule ElixirGoogleScraperWeb.KeywordController do
   def create(conn, %{"file" => %Plug.Upload{} = file}) do
     {alert_type, alert_message} =
       case Scraper.save_keywords(file, conn.assigns.current_user) do
-        {:ok, :file_is_proccessed} -> {:info, "The keyword file is processed successfully!"}
-        {:error, :file_is_empty} -> {:error, "The keyword file is empty!"}
-        {:error, :file_is_invalid} -> {:error, "The keyword file is invalid!"}
-      end
+        {:ok, :file_is_proccessed} ->
+          {:info, "Your CSV file has been uploaded successfully"}
 
-    IO.inspect(alert_type, label: "Alert type")
-    IO.inspect(alert_message, label: "Alert message")
+        {:ok, :file_size_exceed} ->
+          {:info, "File size can't be more than 5 MB."}
+
+        {:ok, :keyword_list_exceed} ->
+          {:info, "CSV Keywords count can't be more than 1000"}
+
+        {:error, :file_is_empty} ->
+          {:error, "File can't be blank"}
+
+        {:error, :file_is_invalid} ->
+          {:error, "The keyword file is invalid!"}
+      end
 
     conn
     |> put_flash(alert_type, alert_message)
